Add simplified version of time.LoadLocation.
Actually it's two functions: GetData and containsDotDot.
The first uses the second to perform a basic sanity name
on location name and then attempts to read the corresponding
timezone file. GetData is a simplified version of LoadLocation.
GetData assumes that it is running on Linux machines and that
the input parameter is sanitized.